+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - add new scope "instruction" for matching mnemonics and operands #767 @williballenthin
  - add new feature "operand[{0, 1, 2}].number" for matching instruction operand immediate values #767 @williballenthin
  - add new feature "operand[{0, 1, 2}].offset" for matching instruction operand offsets #767 @williballenthin
+ - extract additional offset/number features in certain circumstances #320 @williballenthin
 
 ### Breaking Changes